requirements.txt
Django==1.7.1
WebTest==2.0.16
django-webtest==1.7.7
coverage==3.7.1
django-taggit==0.12.2
